editTracksAlbum

+++ src/main/java/dev/MusicSpring/db/repositories/AlbumRepo.java
@@ -1,6 +1,7 @@
 package dev.MusicSpring.db.repositories;
 
 import dev.MusicSpring.db.entities.entity.AlbumEntity;
+import dev.MusicSpring.db.entities.entity.TrackEntity;
 import dev.MusicSpring.db.entities.entity.UserEntity;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
@@ -9,6 +10,6 @@ import java.util.Set;
 
 public interface AlbumRepo extends PagingAndSortingRepository<AlbumEntity, Long> {
     public Set<AlbumEntity> findAll();
-    Optional<AlbumEntity> findById(int id);
-
+    Optional<AlbumEntity> findById(Long id);
+//    Optional<AlbumEntity> findByTrack(TrackEntity track);
 }
